requirements.txt
langchain
openai
PyPDF2
faiss-cpu
tiktoken
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-search-results